Extract OpenBsd/FreeBsd HWDiskStore common bases (JNA/FFM dedup)
